feat(apikeys): adds ClaimrClient class to hold onto API key
The ClaimR API now uses the Authorization header to receive the API key. This is enforced to be set
when using the API for location verification purposes. This commit allows the user to construct a
ClaimR Client object to pass on the API key when a query is made against the ClaimR API.

BREAKING CHANGE: Adds properties to `useLazyVerifiedLocation`, which is an object with field
`client` a `ClaimrClient` object.

Fixes #9