feat: replaces bufferSize with maxEpochs
Replaces the buffer-based approach to storing GNSS data to be epoch aware. Therefore we
are now instead allowing the user to set the amoount of epochs of data to collect,
rather than the total amount of measurements.

This means that we will always send complete epochs, rather than splitting them arbitrarily.

BREAKING CHANGE: The bufferSize option is now removed in the useLazyVerifiedLocation hook.
Instead, use the maxEpochs option. As a rule of thumb, divide the value you were using for
bufferSize by approximately 8 when determining the value for maxEpochs.